+++ tests/test_vcs.py
@@ -20,7 +20,8 @@ class TestVCS(unittest.TestCase):
     def test_git_clone(self):
         vcs.git_clone('https://github.com/audreyr/cookiecutter-pypackage.git')
         self.assertTrue(os.path.isfile('cookiecutter-pypackage/README.rst'))
-        shutil.rmtree('cookiecutter-pypackage')
+        if os.path.isdir('cookiecutter-pypackage'):
+            shutil.rmtree('cookiecutter-pypackage')
 
 if __name__ == '__main__':
     unittest.main()
